[LocalPod] Fixing clean paths
The earlier fix for clean paths is forcing some directories to lower
case which could cause some side effects. This fix removes the
`File::FNM_CASEFOLD` flag from the call to `Dir.glob` and allows the
OS to treat the directories as they wish. This makes the paths
returned in the `clean_paths` method consistent with those returned in
the `expanded_paths` method.