Add Pods integration library
Adds the libPods.a and target derivatives back as targets in the Pods 
project.  The integration libraries are simple containers linked with 
all of the per spec libraries to create a single library for the user 
project to link with.  The xcconfig file for the integration library 
merges namespaced config settings for all dependent specs for use by 
the user project.