[pod_target] Ask the targdef about config inclusion
https://github.com/lookback/Core/commit/affbeae85bd179c5e9a9a927a80f2aacc4d7cf7c
defines a new ":configurations" option for a pod, and exposes the corresponding
API. Using that API in include_in_build_config now lets aggregate_xcconfig
only include the build flags for the libraries that are used.

Next step is to make CocoaPods link using build flags, so that linking
can be disabled with an xcconfig.