Properly check to see if changes were made on push
The original implementation made poor assumptions about the name of the
pod, and used it as a regex pattern. This led to problems when trying to
push specs that contained special characters in their names (such as `+`
for category specs). Instead of using a regex, we should just do a
substring match to see if the output contains the spec name.