[Integration] Fix integration while project is open
While a user project is open and we delete/change the xcconfig files which
the user project references. It causes any subsequent builds to fail
until the project is closed/re-opened, or the project file is touched.

This commit touches the file if we do not need to save it, thus causing
Xcode to reload it.

Closes #2627
Closes #2665