Namespace the pod target product
Pod target products are named after their specs. The namespacing cannot directly happen in the product name itself, because this must be equal to the module name and this will be used in source code, which should stay agnostic over the dependency manager. We need namespacing at all because multiple targets can exist for the same podspec and their products should not collide. This happens when mulitple user targets require the same pod, because they could require different sets of subspecs.