Weak link the aggregate framework
The aggregate target's product isn't really needed, so it isn't copied into the app bundle to not clutter and expose details of the build process. `dyld` is trying to link it at runtime, as the aggregate framework is linked to the user's target to trick Xcode, so that it concludes there is target dependency without the need to declare a literal target dependency, which implies that we would need to add a subproject and add the reference to a target, which is newly created on each run of `pod install`, which would cause again undesirably changes to the user's project.