Fix PODS_ROOT for aggregate target
The value was resolved to 'Pods/Pods', before but correct for the user's target. Because the PODS_ROOT build setting is defined in the aggregate xcconfig, which is shared between the aggregate target and the user's target, we have to overwrite the value in the embedded settings in the Pods project's aggregate target.