Fix duplicate symbol errors by mixed header imports
With static libraries and CocoaPods `#import "…"` and `#import <…>` could be used to import the local headers in "$PODS_ROOT/Headers/…". With frameworks the compiler automatically discovers public headers of the linked frameworks, so we have to ensure that we pass the `-iquote` flag for each framework pointing to its embedded headers. Otherwise this will led to errors because either the `#import "…" won't be discovered at all or they will be seen as different and the compiler will fail with duplicate symbols.