[EmbedFrameworksScript] Ensure that the swift runtime libraries are copied to the application target
This is necessary because they are not copied if the main application target itself does not use swift. Using EMBEDDED_CONTENT_CONTAINS_SWIFT was not used because that would cause the runtime, which is in the realm of 10-20MB, to be copied multiple times, once for each framework that uses swift.