Added specs and fixed duplicate headers
Fixed issue caused by the original fix for #3161, where build and
public headers were included in the Public/Headers/SpecName directory,
even if they were correctly included also in the
Public/Headers/SpecName/FrameworkName directory as the fix for #3161
was addressing. The headers defined in either source_files and
public_header_files will be ignored if they are already included by the
vendored_frameworks.

- Updated banana-lib.tar.gz to include a SubDir/SubBananalib.h in the
Bananalib.framework.

- Updated broken specs because of the new SubDir/SubBananalib.h header,
and also expanded the new vendor spec added for #3161 to include a
validation for the pretense of SubDir/SubBananalib.h.