Add a command to check if dependencies are up to date
Add a command `pod check` that returns success when the
pods specified in the lockfile are installed. This command
is designed to mirror the `bundle check` command provided
by bundler.

The expected use-case is to prevent having to run `pod install`
on every build in a CI scenario. The user should be able to
run `pod check || pod install` rather than pod install. This
can save minutes off CI builds.

http://bundler.io/v1.1/bundle_check.html