Provides a meaningful error message for the dependency conflict occuring when there is no stable version available for a given pod and user do not specify pod version explicitly in Podfile.
While handling resolver error in `handle_resolver_error` method of `resolver.rb` file, we will check if the specificati
on that is not satisfying given requirement was a pre-release version, and if the requirement was including any explicit version or source. If spec is a pre-release version, and there is no explicit version or source for the requirement, we will print a message directing the user to specify pre-release version explicitly.