[Project] Make separate variant groups for different directory levels
It is possible that a pod lib creator includes multiple assets with the same
name, but different file extensions.
Instead of just putting localized files by name, they are now also grouped by
the path of the directory their .lproj folder is in.

For example the library Foo could include a localized Foo.storyboard as well as
a localized Foo.png
Now two separate variant groups would be created for Foo.storyboard and Foo.png
as long as the two files are in different .lproj bundles in different folders.