[PodTargetInstaller] Preserve header_mappings_dir in umbrella header imports
When preserving the directory structure inside a projects header_mappings_dir,
the file structure also needs to be taken into account when generating the
umbrella header for the Pod target.

Instead of the basename, use the relative path in the header_mappings_dir, if
one is set for the Pod.