Scope all PodTargets by their root spec name
This strategy further differs from what Xcode does and breaks native copy files build phases, but allows complete deduplication. It excludes with the new scoping any name clashes where a pod depends on others, which are built in multiple variants, so that it wouldn't be possible without that to add them to the framework search paths, so that just the right variants of each is visible.