[Cache] Use `git ls-remote` to skip full clones for branch dependencies
This adds the posibility of caches hits for git branch depedencies.
Previously, since pods didn't know the commit the branch points it
always did a full clone of the repo. Using `ls-remote` we can check the
commit the branch points to and reuse cached version of the pods if
available.