Fix "does not update non-git repo" test
The test suggests that the fixture is a non-git repo when in reality it
is not. Seemingly, the `.git` file in there references a relative path,
but newer versions of git seem to make it absolute when cloning the
fixtures. This results in the copied repo being treated as a git repo
and the test failing. The fix just removes `.git` after copying, making
the fixture an actual non-git repo :)