generator: prevent header from importing UIKit on iOS & tvOS
This commit replaces the UIKit import made in the header generator
by a Foundation import when targetting iOS and tvOS.

This is done because we don't know if the framework that is built
depends or not on UIKit. For example, frameworks that do not use UI
do not need to import UIKit.

Issue: #6815