[Resolver] Sort external source dependencies first
This resolves a (potential) ordering bug due to pre-releases.

In #possibility_versions_for_root_name, we check whether dependencies on a pod are pre-release or external source, and if so allow pre-releases. It is crucial, therefore, that any external source dependencies have been addressed before any transitive source without the external source, so that pre-releases are (properly) allowed.