Fixes the double .xcasset folder compilation issue
The old code compares a **lowercase** part of a string ( of`$line`) with an uppercase part of a string (of `${PODS_ROOT}`)?
My dev directory lives in `/Users/<username>/Documents/...`.

The `find` command seems to ignore the **uppercase** `Documents` and prints it **lowercase**.
The problem is `$POD_ROOT` contains `Documents` as **uppercase**.

Thus `[[ $line != "${PODS_ROOT}*" ]]` can never be true in my case, am I right?

These changes fix the problem. Any suggestions?