Small refactoring of pod project generator.
Functions in this class now rely less on the assumption that a project object will be created by the time the function is ran and requires a `project` parameter to be passed into itself. Since the `generate!` function is responsible for actually creating the Pods project file, we return it wrapped inside a `PodsProjectGeneratorResult` object alongside the `target_installation_results` that is also returned from `generate!`.