Move calling link headers for all pod targets earlier during installation.
The motivation behind this change is to update the state of the sandbox before project generation so that we can analyze and compare properties on pod targets that depend on the state of the sandbox. For example, trying to grab the `HEADER_SEARCH_PATHS` build settings before linking headers into the sandbox will not include any dependent headers.