Set the path of development pod groups to the top-most shared directory
This fixes an issue that prevents new groups from being placed in the correct directory if all source files are in a nested folder.
For example, if all source code is in `MyLib/Source` directory, then new files or folders should be added within the `Source` folder
instead of the `MyLib` folder.