Update project cache analyzer.
In addition to checking each target's cache key to determine if it's dirty, we should also check the state of its sandbox to protect against a possible case where the user modifies the sandbox and moves or deletes directories.

This change also moves `SandboxDirCleaner` before updating the project cache and Podfile.lock since it makes more sense to write to the project cache and Podfile.lock after installing and cleaning up the sandbox, instead of after.