Stabilize Product Reference UUIDs.
In addition to `PBXNativeTarget` UUIDs being referenced via `remoteGlobalIDString` from other projects, so can the `PBXFileReference` that points to the binaries under the `Products` group. If these UUIDs are not deterministic, then during incremental installation, the UUID for the file reference that pointed to the binary can be changed, resulting in parent project crashing because its `remoteGlobalIDString` might end up pointing to a different object entirely.