Fix regenerating all aggregate targets if one is removed.
This fixes a bug during incremental installation where if an aggregate target is removed, we will not generate `Pods.xcodeproj` since the `AggregateTarget` instance doesn't exist and thus will produce a nil value in an array thereby making it empty. A result of this getting shipped was because the aggregate target test fixtures weren't setup correctly and shared the same target definition in the cache. Verified that the fixed test fails with the old code.