[Analyzer] Do not implicitly whitelist transitive dependencies
Closes #6838

This resolves an issue which caused the transitive dependencies of a dependency with a whitelist (ex. `:configurations => ['Debug']`)
to be linked with the host target even for configurations not included in the whitelist.

This should also provide a small build time improvement for projects using the new build system that contain these dependencies,
when using recent versions of Xcode that detect implicit dependencies based on linker flags (Xcode 10.2 or higher).