Add user projects to workspace during incremental install.
During incremental installation, if the set of `targets_to_generate` was empty and the workspace was missing, we wouldn't add the user projects to the workspace and be missing.

This changes `user_project_paths_to_integrate` to map over `targets` instead of the subset `targets_to_generate` to ensure all user projects will be added. There is existing logic to make sure that duplicate file references to user projects are removed.