Consider `swift_versions` attribute for detecting if a pod uses Swift, only when the pod contains no source code
In the original PR: https://github.com/CocoaPods/CocoaPods/pull/8762
Dimitris raied a concern that "s.swift_versions DSL is currently a root attribute, This means a pod target will be considered as if its using Swift across all of its subspecs whereas some subspecs or tests specs or app specs might not."

What this does instead is that the use of swift_version strictly applies to a pod that has no source code for itself as well its subspecs (usually a header + .a file for framework for example). Should_build is false and therefore the whole pod target can be treated as a single item. In this case the best way to deduce whether it uses Swift or not is through the swift_version flag