Add --configuration options to repo push
This is required to allow pod repo push to succeed with Xcode 12 for specs that
depend on third-party binary only specs. Without this option, the repo push will
fail trying to link arm64 slice built for device to an App built for simulator.

This is one similar lines to --configuraiton option for pod lib lint and pod spec lint