Fix Archive failures when using header_mappings_dir and Catalyst
Catalyst uses the macOS-style versioned framework layout, so it needs the same
workaround for header_mappings_dir as macOS frameworks do. We don't know at
install time if a pod will be used with Catalyst, so I made it always add the
build step for iOS targets and adjusted the script to do nothing if it's
running on a non-versioned framework.