Changed behaviour of the reflect_file_system_structure flag
The project structure now matches Xcodes expectations, so that it is possible move files around without creating a mismatch compared to a freshly generated project.
This could break builds, so the change makes it much more convenient to refactor related project (moving files to different pods, or back to a non-shared product).