Make xcframework's slices order stable
Sort the order of the slices parsed from the `.xcframework`'s `Info.plist`'s `AvailableLibraries` array, so that the order of the extracted slices is consistent across multiple runs.

This is especially useful when [generating the `install_xcframeworks` script invocation](https://github.com/CocoaPods/CocoaPods/blob/4a3159879a29d076d16a20b763f61eb96406a6c1/lib/cocoapods/generator/copy_xcframework_script.rb#L166-L176) (used in the `[CP] Copy XCFrameworks` build phase), to make sure this generated shell script is consistent and stable across runs. Without this fix, every time you re-generate a `xcframework` (using `create_xcframework` fastlane action, for example), the order of the slices generated in the `Info.plist` being arbitrary, they might change order (and they do in practice) so when you then run `pod install` again (which will parse the newly-regenerated `.xcframework` and update the script accordingly), the script might have the order of its args switched.

While changing the order of the slices passed to the script doesn't change the outcome—as that order doesn't matter to the command using those list of slices—it is still annoying to have a non-empty `git diff` every time we regenerate a `.xcframework` again to recompile our library. Having the order of the slices parsed from the `Info.plist` (for which Xcode doesn't offer any guarantee regarding the order in which the slices are listed in) be consistent allows the generation of the `install_xcframeworks` scripts to be idempotent and avoid a needless `diff` when the re-generated framework still contains the same list of slices between multiple compilations / re-generations.