Pass the -f option when resolving the path to the symlinked source.
Fixes the issue where archives fail when using Xcode 14.3.

Xcode 14.3 is now using a relative path in its symlink for frameworks. Without the -f flag, this relative path would be evaluated relative to the working directory of the script being executed, instead of relative to the framework symlink itself. With the -f flag, it resolves that relative path and returns the full path to the source.